Many changes
Moved to using oscrypto instead of cryptography module for lockdown.
It is a much smaller dependency.
Once qh3 is moved to 1.0 also, there will be no more dependency on cryptography module.

Copied in the needed bits of opack with the date stuff removed as it isn't used.

Added env vars to specify where the needed binaries are.

Added a Makefile for easily packaging everything up.

Turns out the easy collection of exceptions into __all__ wasn't working.
For now manually listed out the ones needed till it is solved.