Exit cleanly on EOF (^D)
This cleans up the process for reading stdin a little bit, and now detects both 'stop' and EOF.  (Previously we were creating new streamreaders for every readline() call, and dangling objects seemed to be causing issues on shutdown).

This relies in part on concurrent changes in py_cf_external_utun and utunuds to work completely.